imap_data_extractor_api/requirements.txt
# requirements.txt — VERSION DÉFINITIVE NOVEMBRE 2025
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
ldap3==2.9.1
passlib[argon2]>=1.7.4
python-decouple==3.8
python-dotenv>=1.0
requests

# MongoDB avec Beanie (le seul qui marche vraiment en 2025)
beanie>=1.25
motor>=3.5
pydantic[email]>=2.8


django-auth-adfs
msal
requests
django-fernet-fields
celery


google-api-python-client>=2.95.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0

typing>=3.7.4.3


drf-spectacular

daphne
channels
billiard
channels-redis>=4.3,<5.0
channels>=4.2,<5.0
redis>=4.6

